requirements.txt
pytest-selenium
python-dotenv